requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
babel==2.17.0
billiard==4.2.4
boto3==1.42.24
botocore==1.42.24
cbor2==5.8.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
crispy-bootstrap5==2025.6
cryptography==46.0.3
da-vinci==0.5.0
daphne==4.2.1
Django==5.2.7
django-appconf==1.2.0
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-decouple==2.1
django-filter==25.2
django-htmx==1.27.0
django-multi-form-view==2.0.1
django-pgcrypto==3.1.0
django-pymemcache==1.0.0
django-select2==8.4.8
django-storages==1.14.6
django-tables2==2.8.0
django-tables2-column-shifter==2.2.0
django-thumbnails==0.8.0
django-user-agents==0.4.0
django_celery_results==2.6.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jmespath==1.0.1
kombu==5.6.2
msgpack==1.1.2
packaging==25.0
pendulum==3.1.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pymemcache==4.0.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-memcached==1.62
redis==7.1.0
requests==2.32.5
s3transfer==0.16.0
service-identity==24.2.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==202601
ujson==5.11.0
urllib3==2.6.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.1.1
